fix(checker): split duplicate identifier helper tests
